feat: initial DS setup and first component
