+++ .storybook/preview.ts
@@ -1,4 +1,5 @@
 import type { Preview } from "@storybook/react";
+import "../app/app.css";
 
 const preview: Preview = {
   parameters: {
