+++ app/app.css
@@ -1,14 +1,19 @@
+@import url("https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;500;600;700&display=swap");
 @import "tailwindcss";
 
 @theme {
   --font-sans:
-    "Inter", ui-sans-serif, system-ui, sans-serif, "Apple Color Emoji",
+    "Poppins", ui-sans-serif, system-ui, sans-serif, "Apple Color Emoji",
     "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
+  --color-primary: #073742;
+  --color-secondary: #d1f2fd;
+  --color-light: #ebfafe;
+  --color-lighter: #fafafa;
 }
 
 html,
 body {
-  @apply bg-white dark:bg-gray-950;
+  @apply bg-white font-sans dark:bg-gray-950;
 
   @media (prefers-color-scheme: dark) {
     color-scheme: dark;
